v1-162-part1/pyproject.toml
[tool.poetry]
name = "164-rag-for-pdfs-v2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "^2"
tqdm = "^4.66.2"
unstructured = {extras = ["all-docs"], version = "^0.15.14"}
langchain-experimental = "^0.3.2"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.2"
langchain-community = "^0.3.2"
tiktoken = "^0.8.0"
psycopg = "^3.1.18"
pgvector = "^0.2.5"
langchain-core = "^0.3.10"
langchain-huggingface = "^0.1.0"
sentence-transformers = "^3.2.0"
langchain-postgres = "^0.0.12"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
